requirements.txt
greenlet==2.0.1
importlib-metadata==4.8.3
JayDeBeApi==1.2.3
JPype1==1.3.0
python-dotenv==0.20.0
SQLAlchemy==1.4.44
typing_extensions==4.1.1
zipp==3.6.0
